worker/requirements.txt
redis==4.5.0
psycopg2-binary==2.9.6
